Add smallest positive Integer functionality
Merge sort was added for doing the algorithmic complexity O(nlogn). Also
an implementation for linear search and binary search was added. The
implementation which fits more to the problem is the linear search, but
adding the binary search looks pretty good because of an algorithmic
complexity of O(logn).